Crate sl_liner

A readline-like library For more information refer to the README as well as the examples/ directory for a demonstration of how to create and customize a Context.

Re-exports

pub use keymap::*;

pub use prompt::*;

pub use cursor::*;

pub use editor_rules::*;

Modules

cursor

Track current grapheme offset for terminal cursor

editor_rules

Provides DefaultEditorRules for editor behaviors with custom override function. See NewlineRule and WordDivideRule, the traits that compose editor behaviors.

keymap

Interface for Vi and Emacs KeyMaps

prompt

User-defined prompt.

Structs

BasicCompleter

Completer that can be seeded with a list of prefixes..

Buffer

A buffer for text in the line editor.

Context

Primary interface to readline-like functionality

Editor

The core line editor. Displays and provides editing for history and the new buffer.

EmptyCompleter

Completer with no completions

Event

Event has context about the state of the editor and the EventKind and is consumed by Completer

FilenameCompleter

Completer for filenames in the current working_dir

History

Structure encapsulating command history

Enums

Action

A modification performed on a Buffer. These are used for the purpose of undo/redo.

EventKind

Traits

Completer

Type Aliases

ColorClosure