sl_liner

Module editor_rules

Provides DefaultEditorRules for editor behaviors with custom override function. See NewlineRule and WordDivideRule, the traits that compose editor behaviors.

Structs

DefaultEditorRules

Provides default editor behavior and provides custom override function.

DefaultNewlineRule

Struct with default newline behavior

DefaultWordDivideRule

Struct with default word divide behavior

Traits

EditorRules

Trait that implements all editor rule behaviors, a mechanism to allow modification of the library’s behavior at runtime.

NewlineRule

When the client inputs the newline character, ‘\n’, the Editor can decide to not evalute the newline if the provided implementation of this trait returns false. Default prevents evaluate if the last non whitespace character is a backslash.

WordDivideRule

When the Editor is trying to place the term cursor it needs to know how to divide the words to determine its [CursorPosition] state.

Functions

divide_words_by_space

Default WordDivideRule implementation.

last_non_ws_char_was_not_backslash

Default NewlineRule implementation.