Trait sl_liner::editor_rules::NewlineRule

pub trait NewlineRule {
    // Provided method
    fn evaluate_on_newline(&self, buf: &Buffer) -> bool { ... }
}

When the client inputs the newline character, ‘\n’, the Editor can decide to not evalute the newline if the provided implementation of this trait returns false. Default prevents evaluate if the last non whitespace character is a backslash.

Provided Methods

fn evaluate_on_newline(&self, buf: &Buffer) -> bool