sl_liner::editor_rules

Trait WordDivideRule

pub trait WordDivideRule {
    // Provided method
    fn divide_words(&self, buf: &Buffer) -> Vec<(usize, usize)> { ... }
}

When the Editor is trying to place the term cursor it needs to know how to divide the words to determine its [CursorPosition] state.

Provided Methods

fn divide_words(&self, buf: &Buffer) -> Vec<(usize, usize)>